refactor: 减小custom action体量，重构为pipeline
