+++ .gitignore
@@ -467,3 +467,4 @@ setup_gui.ps1
 setup_gui.bat
 test/
 output/
+Tetris/
