+++ agent/custom/action/__init__.py
@@ -3,9 +3,9 @@ from .AutoFish.auto_fish import *
 from .AutoFish.auto_buy_fish_bait import *
 from .AutoFish.auto_sell_fish import *
 from .auto_make_coffee import *
-from .auto_rhythm import *
+from .rhythm.feats.play import *
+from .rhythm.feats.repeat_decision import *
 from .rhythm.feats.select_song import *
-from .rhythm.feats.repeat import *
 from .Common.click import *
 from .realtime_task import *
 
@@ -17,8 +17,8 @@ __all__ = [
     "AutoBuyFishBait",
     "AutoSellFish",
     "ClickOverride",
-    "AutoRhythm",
+    "AutoRhythmPlay",
+    "AutoRhythmRepeatDecision",
     "AutoRhythmSelectSong",
-    "AutoRhythmRepeat",
     "RealTimeTaskAction",
 ]
