+++ agent/custom/action/rhythm/__init__.py
@@ -1,7 +1,20 @@
 from .utils.lanes import LaneLayout, build_lane_layout
 from .utils.detector import DrumDetector
-from .utils.presence import SceneGate, STATE_OTHER, STATE_SONG_SELECT, STATE_PLAYING, STATE_RESULTS
+from .utils.presence import (
+    SceneGate,
+    STATE_OTHER,
+    STATE_SONG_SELECT,
+    STATE_PLAYING,
+    STATE_RESULTS,
+)
 from .utils.song_selector import SongSelector
-from .utils.assets import list_scene_templates, list_song_templates, list_drum_templates, read_image
+from .utils.assets import (
+    list_scene_templates,
+    list_song_templates,
+    list_drum_templates,
+    read_image,
+)
+from .utils.config import load_rhythm_config
 from .feats.select_song import AutoRhythmSelectSong
-from .feats.repeat import AutoRhythmRepeat
+from .feats.play import AutoRhythmPlay
+from .feats.repeat_decision import AutoRhythmRepeatDecision
