+++ agent/custom/action/rhythm/feats/__init__.py
@@ -1,2 +1,3 @@
 from .select_song import AutoRhythmSelectSong
-from .repeat import AutoRhythmRepeat
+from .play import AutoRhythmPlay
+from .repeat_decision import AutoRhythmRepeatDecision
