+++ agent/custom/action/rhythm/utils/__init__.py
@@ -3,3 +3,4 @@ from .detector import DrumDetector
 from .presence import SceneGate, STATE_OTHER, STATE_SONG_SELECT, STATE_PLAYING, STATE_RESULTS
 from .song_selector import SongSelector
 from .assets import list_scene_templates, list_song_templates, list_drum_templates, read_image
+from .config import load_rhythm_config
