+++ assets/resource/base/pipeline/Rhythm/Rhythm.json
@@ -3,36 +3,6 @@
         "desc": "演奏功能入口",
         "action": "DoNothing",
         "next": [
-            "RhythmSetSong"
-        ],
-        "on_error": [
-            "RhythmExit"
-        ]
-    },
-    "RhythmSetSong": {
-        "desc": "[内部] 记录演奏目标歌曲",
-        "action": "DoNothing",
-        "next": [
-            "RhythmSetRepeat"
-        ],
-        "on_error": [
-            "RhythmExit"
-        ]
-    },
-    "RhythmSetRepeat": {
-        "desc": "[内部] 记录连打次数",
-        "action": "DoNothing",
-        "next": [
-            "RhythmSetFPS"
-        ],
-        "on_error": [
-            "RhythmExit"
-        ]
-    },
-    "RhythmSetFPS": {
-        "desc": "[内部] 记录目标帧率",
-        "action": "DoNothing",
-        "next": [
             "RhythmSelectSong"
         ],
         "on_error": [
@@ -45,16 +15,86 @@
         "custom_action": "auto_rhythm_select_song",
         "custom_action_param": {},
         "next": [
-            "RhythmStart"
+            "RhythmWaitPlaying"
         ],
         "on_error": [
-            "RhythmStart"
+            "RhythmWaitPlaying"
         ]
     },
-    "RhythmStart": {
-        "desc": "演奏主流程",
+    "RhythmWaitPlaying": {
+        "desc": "等待进入演奏界面",
+        "recognition": "And",
+        "all_of": [
+            "RhythmSceneOnPlaying"
+        ],
+        "pre_wait_freezes": {
+            "target": [0, 0, 100, 100],
+            "time": 500,
+            "timeout": 10000
+        },
+        "next": [
+            "RhythmPlaying"
+        ]
+    },
+    "RhythmPlaying": {
+        "desc": "演奏主流程 - 鼓面检测紧凑循环",
+        "anchor": "RhythmLoopPoint",
+        "max_hit": 999,
         "action": "Custom",
-        "custom_action": "auto_rhythm",
+        "custom_action": "auto_rhythm_play",
+        "custom_action_param": {},
+        "next": [
+            "RhythmAfterPlaying"
+        ],
+        "on_error": [
+            "RhythmExit"
+        ]
+    },
+    "RhythmAfterPlaying": {
+        "desc": "演奏结束，判断当前状态",
+        "next": [
+            "RhythmOnResults",
+            "RhythmOnSongSelect",
+            "[Anchor]RhythmLoopPoint"
+        ]
+    },
+    "RhythmOnResults": {
+        "desc": "检测到结算界面",
+        "recognition": "And",
+        "all_of": [
+            "RhythmSceneOnResults"
+        ],
+        "next": [
+            "RhythmDismissResults"
+        ]
+    },
+    "RhythmOnSongSelect": {
+        "desc": "检测到选歌界面",
+        "recognition": "And",
+        "all_of": [
+            "RhythmSceneOnSongSelect"
+        ],
+        "next": [
+            "RhythmSelectSong"
+        ]
+    },
+    "RhythmDismissResults": {
+        "desc": "退出结算界面",
+        "action": {
+            "type": "ClickKey",
+            "param": {
+                "key": [27]
+            }
+        },
+        "post_delay": 800,
+        "next": [
+            "RhythmRepeatCheck"
+        ]
+    },
+    "RhythmRepeatCheck": {
+        "desc": "连打决策",
+        "action": "Custom",
+        "custom_action": "auto_rhythm_repeat_decision",
         "custom_action_param": {},
         "next": [
             "RhythmExit"
